Refactor MemedFactory, MemedToken, and MemedWarriorNFT contracts to utilize new interface implementations and improve code organization. This update replaces direct contract references with interfaces, enhancing modularity and maintainability while streamlining the token creation process.

+++ interfaces/IMemedBattle.sol
@@ -14,15 +14,15 @@ interface IMemedBattle {
     function tokenAllocations(
         address _user
     ) external view returns (uint256[] memory);
-    function getBattle(uint256 _battleId) external view returns (Battle memory);
+    function getBattle(uint128 _battleId) external view returns (Battle memory);
     function getBattleAllocations(
-        uint256 _battleId,
+        uint128 _battleId,
         address _user,
         address _meme
     ) external view returns (UserBattleAllocation memory);
-    function getBackWarrior(uint256 _battleId, address _user) external;
+    function getBackWarrior(uint128 _battleId, address _user) external;
     function allocateNFTsToBattle(
-        uint256 _battleId,
+        uint128 _battleId,
         address _user,
         address _supportedMeme,
         uint256[] calldata _nftsIds
